+++ src/imio/restapi/serializer/base.py
@@ -1,14 +1,19 @@
 # -*- coding: utf-8 -*-
 
 from Acquisition import aq_inner
-from plone.restapi.serializer import dxcontent
-from plone.restapi.interfaces import ISerializeToJson
-from zope.component import adapter
-from zope.interface import implementer
+from imio.restapi.interfaces import IImioRestapiLayer
+from imio.restapi.utils import listify
+from Products.ZCatalog.interfaces import ICatalogBrain
+from plone import api
 from plone.dexterity.interfaces import IDexterityContainer
 from plone.dexterity.interfaces import IDexterityContent
-from imio.restapi.interfaces import IImioRestapiLayer
-from plone import api
+from plone.restapi.interfaces import ISerializeToJson
+from plone.restapi.interfaces import ISerializeToJsonSummary
+from plone.restapi.serializer import dxcontent
+from plone.restapi.serializer import summary
+from zope.component import adapter
+from zope.interface import implementer
+from zope.interface import Interface
 
 
 @implementer(ISerializeToJson)
@@ -35,3 +40,28 @@ def get_relative_path(context):
 
     relative_path = context.getPhysicalPath()[len(portal.getPhysicalPath()):]
     return "/{}".format("/".join(relative_path))
+
+
+@implementer(ISerializeToJsonSummary)
+@adapter(ICatalogBrain, Interface)
+class DefaultJSONSummarySerializer(summary.DefaultJSONSummarySerializer):
+    """Formalize management of defining extra metadata_fields in the serializer."""
+
+    @property
+    def _additional_fields(self):
+        """By default add 'id' and 'UID' to returned data."""
+        return ["id", "UID"]
+
+    def _set_metadata_fields(self):
+        """Must be set in request.form."""
+        form = self.request.form
+        # manage metadata_fields
+        additional_metadata_fields = listify(form.get("metadata_fields", []))
+        additional_metadata_fields += self._additional_fields
+        form["metadata_fields"] = additional_metadata_fields
+
+    def __call__(self):
+        """ """
+        self._set_metadata_fields()
+        result = super(DefaultJSONSummarySerializer, self).__call__()
+        return result